Further bug fixes due to  Rcpp error for Webb weights with few clusters #11 (raised by fschoner).
For rademacher and mammen weights, if 2^(number of clusters) < B, the if-clause that should initiate full enumeration was not binding - in consequence, not every feasible t-statistics was generated & some were (randomly) oversampled with B = 2^(number of clusters), leading to theoretically impossible p-values for (deterministic) full enumeration.

Unit tests for correct behavior are added in test_seed and test_small_n_cluster.

Note: for the case of few clusters, p-value inversion to calculate standard errors might often not work (as no initial p-values can be found). TO DO: clarify this in the resulting error message & check how stata::boottest() handles such cases.