nudge citation to fwildclusterboot
I noticed that Mark has (rightly) more than 1k citations to stargazer. I think one of the reasons is that the package produces a message onAttach to remind the user to cite the package. I have made a similar change here. I think there is a larger point here about increasing citations to software to improve incentives for software producers and also make sure concerns in software can be backtraced from papers that rely on it.